+++ app/src/main/java/com/philipgreen/dmwizard/playerClasses/Barbarian.java
@@ -4,7 +4,7 @@ import android.util.Log;
 
 import com.philipgreen.dmwizard.data.ArmorProficiencies;
 import com.philipgreen.dmwizard.data.BaseStats;
-import com.philipgreen.dmwizard.data.Skills;
+import com.philipgreen.dmwizard.player.utils.Skills;
 import com.philipgreen.dmwizard.data.WeaponProficiencies;
 
 import java.util.Arrays;
